requirements.txt
flask
flask-wtf
flask-bootstrap
wtform
django
flask-httpauth
flask-login